terraform/jsonfeed/requirements.txt
google-cloud-bigquery
flask
